requirements.txt
wheel
numpy
cmake
farm-haystack
pypdf 
openai